Changed Order to int and formatted using the proper style

+++ WebActivator/BaseActivationMethodAttribute.cs
@@ -25,7 +25,7 @@ namespace WebActivator {
             }
         }
 
-        public byte Order { get; set; }
+        public int Order { get; set; }
 
 
         public void InvokeMethod() {
